basic-shopping-cart/auth/requirements.txt
python-dotenv
Falcon
requests
psycopg2